Avoid deleting not found index records.
This can be the case when data is deleted that was not index before for some reason.

+++ src/Model/Behavior/ElasticIndexBehavior.php
@@ -156,6 +156,9 @@ class ElasticIndexBehavior extends Behavior {
     public function deleteIndexDocument(EntityInterface $entity)
     {
         $elasticEntity = $this->_findElasticDocument($entity);
+        if (empty($elasticEntity)) {
+            return;
+        }
         $this->elasticIndex()->delete($elasticEntity);
     }
